일일 파일 전송량 표시 갱신 되도록 기능 구현.  - 사용한 일일 파일 전송량을 가져와 남은 파일 전송가능 용량 및 횟수 를 구하는 함수 구현.  - 사용한 일일 파일 전송량에 대한 % 계산 함수 구현.  - 서버에서 수신 받은 데이터 파일 전송화면과 연동.

+++ src/OpenNetLinkApp/Services/PageStatusService.cs
@@ -307,7 +307,7 @@ namespace OpenNetLinkApp.Services
             return m_DicPageStatusData[groupID].GetDayRemainClipCount();
         }
 
-        public int GetDayRemainFileSizePercent(int groupID)
+        public double GetDayRemainFileSizePercent(int groupID)
         {
             PageStatusData tmpData = null;
             if (m_DicPageStatusData.TryGetValue(groupID, out tmpData) != true)
@@ -317,7 +317,7 @@ namespace OpenNetLinkApp.Services
             return m_DicPageStatusData[groupID].GetDayRemainFileSizePercent();
         }
 
-        public int GetDayRemainFileCountPercent(int groupID)
+        public double GetDayRemainFileCountPercent(int groupID)
         {
             PageStatusData tmpData = null;
             if (m_DicPageStatusData.TryGetValue(groupID, out tmpData) != true)
@@ -327,7 +327,7 @@ namespace OpenNetLinkApp.Services
             return m_DicPageStatusData[groupID].GetDayRemainFileCountPercent();
         }
 
-        public int GetDayRemainClipSizePercent(int groupID)
+        public double GetDayRemainClipSizePercent(int groupID)
         {
             PageStatusData tmpData = null;
             if (m_DicPageStatusData.TryGetValue(groupID, out tmpData) != true)
@@ -337,7 +337,7 @@ namespace OpenNetLinkApp.Services
             return m_DicPageStatusData[groupID].GetDayRemainClipSizePercent();
         }
 
-        public int GetDayRemainClipCountPercent(int groupID)
+        public double GetDayRemainClipCountPercent(int groupID)
         {
             PageStatusData tmpData = null;
             if (m_DicPageStatusData.TryGetValue(groupID, out tmpData) != true)
